Ilya_Efimov/lesson 3/yt-intents-agent/requirements.txt
openai
mcp[cli]
rapidfuzz
python-dotenv
requests
python-telegram-bot